+++ photon-model-tasks/src/main/java/com/vmware/photon/controller/model/tasks/EndpointAllocationTaskService.java
@@ -238,7 +238,7 @@ public class EndpointAllocationTaskService
         patchBody.taskInfo.stage = TaskStage.STARTED;
         patchBody.taskSubStage = nextStage;
 
-        ComputeSubTaskService.ComputeSubTaskState subTaskInitState = new ComputeSubTaskService.ComputeSubTaskState();
+        SubTaskService.SubTaskState subTaskInitState = new SubTaskService.SubTaskState();
         subTaskInitState.parentPatchBody = Utils.toJson(patchBody);
         subTaskInitState.errorThreshold = 0;
         subTaskInitState.parentTaskLink = getSelfLink();
@@ -247,7 +247,7 @@ public class EndpointAllocationTaskService
         Operation startPost = Operation
                 .createPost(this, UUID.randomUUID().toString())
                 .setBody(subTaskInitState).setCompletion(c);
-        getHost().startService(startPost, new ComputeSubTaskService());
+        getHost().startService(startPost, new SubTaskService());
     }
 
     private void sendEnhanceRequest(Object body, EndpointAllocationTaskState currentState) {